add existence check for identical message in webhook_calls before processing

+++ src/ProcessSesWebhookJob.php
@@ -21,6 +21,10 @@ class ProcessSesWebhookJob extends ProcessWebhookJob
 
     public function handle()
     {
+        if ($this->webhookCall->identicalMessageExists()) {
+            return;
+        }
+
         if (! $message = $this->getMessageFromWebhookCall()) {
             $this->webhookCall->delete();
